requirements.txt
openai>=1.43.0
langchain>=0.2.12
langchain-openai>=0.1.22
langchain-community>=0.2.11
qdrant-client>=1.9.2
python-dotenv>=1.0.1
pydantic>=2.8.2
typer>=0.12.3
rich>=13.7.1
tabulate>=0.9.0
pytest>=8.3.2
